backend/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.7.0

# HTTP client
httpx>=0.28.0

# MCP
mcp>=1.27.0

# LLM providers
google-genai>=1.14.0
anthropic>=0.45.0
openai>=1.60.0

# Database
asyncpg>=0.30.0

# Charts
matplotlib>=3.10.0

# Security - PII
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
spacy>=3.8.0

# Security - Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Security - Rate limiting
slowapi>=0.1.9

# Security - SQL parsing
sqlparse>=0.5.0

# Email
aiosmtplib>=3.0.0

# Logging
structlog>=25.1.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.25.0
